Add link to authorized offical, if available
This change attempts to find which provider, if any, match the authorized official to the organization and provides a link next to the officials name on the organization page. Since this is an estimate, and multiple users may fit the match (check first and last name for matching) it can produce multiple, numerically identifiable links next to the name.

For example, if John Smith had two mentions for one organization, the result would appear as "John Smith [0] [1]".